Allow to drop Calcite Core and Avatica dependecies and fix MySQL Backticks
